+++ bot/src/economy/index.ts
@@ -3,5 +3,6 @@ export * from "./context.js";
 export * from "./format.js";
 export * from "./orders.js";
 export * from "./parse.js";
+export * from "./sc-craft.js";
 export * from "./service.js";
 export * from "./uex.js";